make issue table container into individual class. for let other issue-like table to inherit.
